pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Project: Difference Calculator"
authors = [{ name = "RStructura" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
